+++ pfrl/nn/constants_mult.py
@@ -1,0 +1,10 @@
+from torch import nn
+
+
+class ConstantsMult(nn.Module):
+    def __init__(self, constants):
+        super().__init__()
+        self.constants = constants
+
+    def forward(self, x):
+        return self.constants * x
